// apps/web/react/app/src/generated/api/models/ArchivedTask.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ArchivedTask = {
    archivedAt: string;
    boardId: string;
    columnId: string;
    createdAt: string;
    description: string;
    id: string;
    position: number;
    title: string;
    updatedAt: string;
};