// apps/web/react/app/src/generated/api/models/CreateTaskRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateTaskRequest = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    columnId: string;
    description: string;
    title: string;
};