// apps/web/react/app/src/generated/api/models/Task.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Task = {
    /**
     * A URL to the JSON Schema for this object.
     */
    readonly $schema?: string;
    boardId: string;
    columnId: string;
    createdAt: string;
    description: string;
    id: string;
    position: number;
    title: string;
    updatedAt: string;
};